requirements.txt
click==8.2.1
iniconfig==2.1.0
markdown-it-py==4.0.0
mdurl==0.1.2
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
python-dotenv==1.1.1
PyYAML==6.0.2
rich==14.1.0
shellingham==1.5.4
typer==0.16.1
typing_extensions==4.14.1
pandas
openpyxl
requests
